refactor: adds FileSystemBreadcrumbs component
